Renamed the core. Updated the dipendencies.
