+++ Sandbox.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31624.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{9444F20F-8353-494F-8067-AC32EC0F967F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SandboxCore", "SandboxCore\SandboxCore.csproj", "{9444F20F-8353-494F-8067-AC32EC0F967F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SandboxGh", "SandboxGh\SandboxGh.csproj", "{2326E4F3-873E-4357-8B70-12771892F58E}"
 EndProject
