+++ SandboxCore/Properties/AssemblyInfo.cs
