+++ SandboxCore/SandboxCore.csproj
@@ -13,7 +13,7 @@
     <RepositoryUrl>https://github.com/mottmacdonaldglobal/Sandbox</RepositoryUrl>
     <PackageTags>computational geometry, architecture, engineering, construction, csharp</PackageTags>
     <PackageDescription>🧰 A library helping to speed repeated and annoying tasks, the library collects tools for 🦗 and 🧨.</PackageDescription>
-    <RootNamespace>Core</RootNamespace>
+    <RootNamespace>SandboxCore</RootNamespace>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -24,13 +24,13 @@
     <When Condition="'$(AssemblyInformationalVersion)' == '' ">
       <PropertyGroup>
         <Version>0.0.0</Version>
-        <AssemblyName>Core</AssemblyName>
+        <AssemblyName>SandboxCore</AssemblyName>
       </PropertyGroup>
     </When>
     <When Condition="'$(AssemblyInformationalVersion)' != '' ">
       <PropertyGroup>
         <Version>$(AssemblyInformationalVersion)</Version>
-        <AssemblyName>Core_$(AssemblyInformationalVersion)</AssemblyName>
+        <AssemblyName>SandboxCore_$(AssemblyInformationalVersion)</AssemblyName>
       </PropertyGroup>
     </When>
   </Choose>
