+++ SandboxCore/SpreadSheetReader.cs
@@ -1,14 +1,14 @@
-﻿using CsvHelper;
-using CsvHelper.Configuration;
-using NPOI.SS.UserModel;
-using NPOI.SS.Util;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using CsvHelper;
+using CsvHelper.Configuration;
+using NPOI.SS.UserModel;
+using NPOI.SS.Util;
 
-namespace Core
+namespace SandboxCore
 {
     public static class SpreadSheetReader
     {
