+++ SandboxDyn/manifests/pkg.json
@@ -1,1 +1,1 @@
-{ "license":"MIT", "file_hash":null, "name":"Sandbox", "version":"1.0.0.0", "description":"A library of tools.", "group":"", "keywords":[ "NPOI", "CSVHelper", "MottMacDonald"], "dependencies":[], "contents":"A library of dynamo noods.", "engine_version":"2.12.0.5650", "engine":"dynamo", "engine_metadata":"", "site_url":"", "repository_url":"https://github.com/mottmacdonaldglobal/Sandbox", "contains_binaries":true,  "node_libraries":[ "Sandbox,  Version=1.0.0.0,  Culture=neutral,  PublicKeyToken=null"]} 
+{ "license":"MIT", "file_hash":null, "name":"SandboxDyn", "version":"0.0.3-alpha.0", "description":"A library of tools.", "group":"", "keywords":[ "NPOI", "CSVHelper", "MottMacDonald"], "dependencies":[], "contents":"A library of dynamo noods.", "engine_version":"2.12.0.5650", "engine":"dynamo", "engine_metadata":"", "site_url":"", "repository_url":"https://github.com/mottmacdonaldglobal/Sandbox", "contains_binaries":true,  "node_libraries":[ "SandboxDyn,  Version=0.0.3-alpha.0,  Culture=neutral,  PublicKeyToken=null"]}
